requirements.txt
asyncpg==0.22.0
fastapi==0.63.0
uvicorn==0.13.4
strawberry-graphql
requests
http3
celery
redis
pika
psycopg2-binary>=2.8
pymongo
pytest